Add application logic to enable automatic time stamp format detection

+++ src/main/java/de/logrifle/data/parsing/SampleContentFetcher.java
@@ -23,5 +23,5 @@ package de.logrifle.data.parsing;
 import java.util.Collection;
 
 public interface SampleContentFetcher {
-    Collection<String> getSampleContent(long desiredLinesCount);
+    Collection<String> getSampleContent(int desiredLinesCount);
 }
